@@ -42,20 +50,25 @@
   "dependencies": {
     "axios": "0.16.2",
     "babel-polyfill": "6.26.0",
+    "fixed-data-table-2": "0.8.1",
     "history": "4.7.2",
+    "hoist-non-react-statics": "2.3.1",
+    "pluralize": "7.0.0",
     "prop-types": "15.5.10",
     "ramda": "0.24.1",
     "ramda-adjunct": "1.15.0",
-    "react": "15.6.1",
+    "react": "16.0.0",
     "react-apollo": "1.4.15",
-    "react-dom": "15.6.1",
+    "react-dom": "16.0.0",
+    "react-fastclick": "3.0.2",
     "react-localization": "0.1.2",
     "react-redux": "5.0.6",
     "react-router-dom": "4.1.2",
     "react-router-redux": "4.0.8",
     "react-scripts": "1.0.13",
-    "react-tap-event-plugin": "2.0.1",
+    "react-tooltip": "3.3.1",
     "redux": "3.7.2",
+    "redux-auth-wrapper": "2.0.2",
     "redux-form": "7.0.4",
     "redux-logger": "3.0.6",
     "redux-saga": "0.15.6",
