@@ -1,7 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import ErrorMessage from '../common/ErrorMessage'
+import {ErrorMessage} from '../common'
+import {getThemeProp} from '../common/helpers'
 
 const ErroMessageWithCloseIcon = styled(ErrorMessage)`
     position: relative;
@@ -11,7 +12,7 @@ const ErroMessageWithCloseIcon = styled(ErrorMessage)`
 
     &:after {
         content: '${'\u2715'}';
-        color: white;
+        color: ${getThemeProp(['colors', 'grayscale', 'white'], 'white')};
         position: absolute;
         top: 4px;
         right: 4px;
