@@ -1,12 +1,19 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Button from '../common/Button'
+import {Button, ButtonLink} from '../common'
 
-const Logout = ({tryLogout}) =>
-    <Button onClick={tryLogout} type="button">Logout</Button>
+const Logout = ({tryLogout, asLink}) => (
+    asLink ? <ButtonLink onClick={tryLogout}>Logout</ButtonLink>
+    : <Button onClick={tryLogout} type="button">Logout</Button>
+)
 
 Logout.propTypes = {
-    tryLogout: PropTypes.func.isRequired
+    tryLogout: PropTypes.func.isRequired,
+    asLink: PropTypes.bool.isRequired
+}
+
+Logout.defaultProps = {
+    asLink: false
 }
 
 export default Logout
