@@ -4,7 +4,7 @@ import {graphql} from 'react-apollo'
 import Validator from 'validatorjs'
 
 import Registration from './Registration'
-import {handleError, cancel, register} from './actions'
+import {handleError, register} from './actions'
 import constants from './constants'
 import {REGISTER_USER} from './mutations'
 
@@ -15,18 +15,6 @@ const validate = values => {
     validator.passes()
     return validator.errors.all()
 }
-
-const mapDispatchToProps = dispatch => ({
-    handleError(error) {
-        return dispatch(handleError(error))
-    },
-    cancel() {
-        return dispatch(cancel())
-    },
-    registerUser(user) {
-        return dispatch(register(user))
-    }
-})
 
 const FormedRegistration = reduxForm({
     validate,
@@ -42,7 +30,7 @@ const RegistrationWithData = graphql(REGISTER_USER, {
                 if (error) {
                     throw new Error(error)
                 }
-                if (registerUser) ownProps.registerUser(user)
+                if (registerUser) ownProps.register(user)
             } catch (err) {
                 ownProps.handleError(err)
             }
@@ -50,4 +38,5 @@ const RegistrationWithData = graphql(REGISTER_USER, {
     })
 })(FormedRegistration)
 
-export default connect(null, mapDispatchToProps)(RegistrationWithData)
+export default connect(null, {handleError, register})(RegistrationWithData)
+
