@@ -2,15 +2,13 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
-import Button from '../common/Button'
-import Form from '../common/Form'
-import Logo from '../common/Logo'
-import ReduxFormField from '../common/ReduxFormField'
-import {breakpoints, forms} from '../common/constants'
+import {Link} from 'react-router-dom'
+import {Button, LinkButton, Form, SimpleSvgIcon, ReduxFormField} from '../common'
+import {getThemeProp} from '../common/helpers'
 
 const StyledForm = styled(Form)`
     & > * {
-        margin: ${forms.formItemMargin};
+        margin: ${getThemeProp(['forms', 'formItemMargin'], '5px')};
     }
 
     & .attainiaLogo {
@@ -35,7 +33,7 @@ const StyledForm = styled(Form)`
     }
 
     @supports (display: grid) {
-        @media ${breakpoints.desktop} {
+        @media ${getThemeProp(['breakpoints', 'tablet'], 'screen and (min-width: 768px)')} {
             display: grid;
             grid-template-columns: 1fr 1fr;
             grid-template-areas: 'header header' 'instructions instructions' 'name name' 'email email'
@@ -71,24 +69,19 @@ const StyledForm = styled(Form)`
         }
     }
 `
-const Registration = ({handleSubmit, tryRegister, cancel}) =>
+const Registration = ({handleSubmit, tryRegister}) =>
     <StyledForm onSubmit={handleSubmit(tryRegister)}>
-        <Logo className="attainiaLogo" />
+        <SimpleSvgIcon className="attainiaLogo" width="161" height="39" icon="primary" />
         <p className="instructions">Register Your Account</p>
         <ReduxFormField className="name" placeholder="name" name="name" />
         <ReduxFormField className="email" placeholder="email" name="email" type="email" />
         <ReduxFormField className="password" placeholder="password" type="password" name="password" />
-        <Button className="registrationButton" type="submit">
-            Register
-        </Button>
-        <Button className="cancelButton" type="button" onClick={cancel}>
-            Cancel
-        </Button>
+        <Button className="registrationButton" type="submit">Register</Button>
+        <LinkButton className="cancelButton"><Link to="/">Cancel</Link></LinkButton>
     </StyledForm>
 
 Registration.propTypes = {
     handleSubmit: PropTypes.func.isRequired,
-    cancel: PropTypes.func.isRequired,
     tryRegister: PropTypes.func.isRequired
 }
 
