@@ -1,12 +1,8 @@
 import types from './types'
 
-export const cancel = () => ({type: types.CANCEL})
 export const handleError = error => ({error, type: types.ERROR})
 export const clearError = () => ({type: types.CLEAR_ERROR})
 export const getUserNavMenu = navigation => ({navigation, type: types.GET_USER_NAV_MENU})
-export const gotoPasswordHelp = () => ({type: types.GOTO_PASSWORD_HELP})
-export const gotoRegistration = () => ({type: types.GOTO_REGISTRATION})
-export const gotoAppRegistration = () => ({type: types.GOTO_APP_REGISTRATION})
 export const passwordHelp = email => ({email, type: types.PASSWORD_HELP})
 export const refresh = refreshTimeout => ({refreshTimeout, type: types.REFRESH})
 export const clearRefresh = () => ({type: types.CLEAR_REFRESH})
@@ -16,3 +12,8 @@ export const toggleRememberMe = rememberMe => ({rememberMe, type: types.REMEMBER
 export const login = user => ({user, type: types.LOGIN})
 export const logout = () => ({type: types.LOGOUT})
 export const updatedToken = token => ({token, type: types.UPDATED_TOKEN})
+export const parsedToken = token => ({token, type: types.PARSED_TOKEN})
+export const validatedToken = token => ({token, type: types.VALIDATED_TOKEN})
+export const userInfoFromToken = user => ({user, type: types.USER_INFO_FROM_TOKEN})
+export const startedLoading = () => ({type: types.LOADING_STARTED})
+export const finishedLoading = () => ({type: types.LOADING_FINISHED})
