@@ -1,13 +1,13 @@
 export default {
     app: {},
+    baseUrl: 'localhost',
     error: '',
-    route: 'login',
-    status: '',
-    refreshStatus: '',
     rememberMe: false,
     menu: {
         navigation: [],
         profile: []
     },
-    user: {}
+    storageType: 'local',
+    user: {},
+    loading: false
 }
