@@ -1,6 +1,6 @@
 import {gql} from 'react-apollo'
 
-export const IS_LOGGED_OUT = gql`
+export default gql`
     subscription isLoggedOut($token: String!) {
         isLoggedOut(token: $token)
     }
