@@ -1,9 +1,9 @@
 import styled from 'styled-components'
-import {colors} from './constants'
+import {getThemeProp} from './helpers'
 
 export default styled.div`
     color: white;
-    background-color: ${colors.rossoCorsa};
+    background-color: ${getThemeProp(['colors', 'status', 'error'], 'crimson')};
     font-size: 10px;
     padding: 7px;
     text-align: center;
