@@ -1,10 +1,10 @@
 import styled from 'styled-components'
-import {colors} from './constants'
+import {getThemeProp} from './helpers'
 
 export default styled.a`
-    color: ${colors.blue};
+    color: ${getThemeProp(['colors', 'secondary', 'dk'], 'royalblue')};
     text-decoration: underline;
-    font-size: 12px;
+    font-size: ${getThemeProp(['fonts', 'fontSize'], '12px')};
     &:focus {
         outline: none;
     }
