@@ -1,13 +1,12 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styled from 'styled-components'
-import Logo from '../common/Logo'
-import ApplicationSearchButton from '../common/ApplicationSearchButton'
-import NotificationButton from '../common/NotificationButton'
-import {colors} from '../common/constants'
+import styled, {withTheme} from 'styled-components'
+import {SimpleSvgIcon} from '../common'
+import {getThemeProp} from '../common/helpers'
+import {LogoutContainer} from '../auth'
 
 const Li = styled.div`
-    color: ${colors.cornflowerBlue};
+    color: ${getThemeProp(['colors', 'secondary', 'default'], 'royalblue')};
     text-align: center;
     background: white;
     height: 100%;
@@ -32,7 +31,7 @@ const ListHeader = styled.ul`
     list-style: none;
     margin: 0;
     padding: 0;
-    background: ${colors.isabellineGray};
+    background: ${getThemeProp(['colors', 'grayscale', 'white'], 'white')};
 
     .logo {
         padding-left: 50px;
@@ -55,20 +54,30 @@ const ListHeader = styled.ul`
         grid-column-gap: 3px;
     }
 `
-const Header = ({className}) =>
-    <ListHeader className={className}>
-        <Li className="logo"><Logo className="headerLogo" /></Li>
+const Header = props =>
+    <ListHeader className={props.className}>
+        <Li className="logo">
+            <SimpleSvgIcon icon="primary" className="headerLogo" />
+        </Li>
         <Li className="btnSearch">
-            <ApplicationSearchButton />
+            <SimpleSvgIcon
+                icon="search"
+                fill={getThemeProp(['colors', 'secondary', 'default'])(props)}
+            />
         </Li>
         <Li className="btnNotifications">
-            <NotificationButton />
+            <SimpleSvgIcon
+                icon="notification"
+                fill={getThemeProp(['colors', 'secondary', 'default'])(props)}
+            />
         </Li>
-        <Li className="profileMenu"><div>User Profile</div></Li>
+        <Li className="profileMenu">
+            <LogoutContainer asLink>Logout</LogoutContainer>
+        </Li>
     </ListHeader>
 
 Header.propTypes = {
     className: PropTypes.string.isRequired
 }
 
-export default Header
+export default withTheme(Header)
