@@ -1,8 +1,19 @@
-import styled from 'styled-components'
-import {breakpoints} from '../common/constants'
+import styled, {keyframes} from 'styled-components'
+import {getThemeProp} from '../common/helpers'
 
+const fadeIn = keyframes`
+    from {
+        opacity: 0.5;
+    }
+
+    to {
+        opacity: 1;
+    }
+`
 export default styled.div`
     height: 100%;
+    animation: ${fadeIn} .5s ease;
+
     @supports not (display: grid) {
         .header,
         .main,
@@ -14,7 +25,7 @@ export default styled.div`
     }
 
     @supports (display: grid) {
-        @media ${breakpoints.tablet} {
+        @media ${getThemeProp(['breakpoints', 'tablet'], 'screen and (min-width: 768px)')} {
             display: grid;
             grid-template-columns: 200px 1fr;
             grid-template-rows: 50px auto 40px;
@@ -27,7 +38,7 @@ export default styled.div`
             .main {
                 grid-area: main;
                 display: grid;
-                align-items: center;
+                align-items: start;
             }
 
             .sidebar {
@@ -39,7 +50,7 @@ export default styled.div`
             }
         }
 
-        @media ${breakpoints.largeDesktop} {
+        @media ${getThemeProp(['breakpoints', 'largeDesktop'], 'screen and (min-width: 1200px)')} {
             display: grid;
             grid-template-columns: 200px 1fr 1fr;
             grid-template-areas: 'header header header' 'sidebar main main' 'footer footer footer';
@@ -50,7 +61,7 @@ export default styled.div`
             .main {
                 grid-area: main;
                 display: grid;
-                align-items: center;
+                align-items: start;
             }
 
             .sidebar {
