@@ -1,4 +1,3 @@
 import auth from './components/auth/types'
-import resources from './components/resources/types'
 
-export default {auth, resources}
+export default {auth}
